Small bugfix, renamed checkard => checkcards
