+++ index.js
@@ -10,7 +10,7 @@ const {
     giftrolls,
     newplayer,
     findplayer,
-    checkcard
+    checkcards
 } = require("./commands/commands.js");
 
 const config = require('./config/config.js');
@@ -59,8 +59,8 @@ client.on('messageCreate', async message => {
         case 'giftrolls':
             responses = responses.concat(await giftrolls(player, message, args));
             break;
-        case 'checkcard':
-            responses = responses.concat(await checkcard(player,message,args));
+        case 'checkcards':
+            responses = responses.concat(await checkcards(player,message,args));
             break;
     }
     show(responses, message);
